+++ relevanceai/api/batch/batch_insert.py
@@ -119,6 +119,8 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
             chunksize=chunksize,
         )
 
+    _insert_documents = insert_documents
+
     def insert_csv(
         self,
         dataset_id: str,
@@ -198,6 +200,8 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
             "failed_documents": failed_documents,
             "failed_documents_detailed": failed_documents_detailed,
         }
+
+    _insert_csv = insert_csv
 
     def _insert_csv_chunk(
         self,
@@ -551,7 +555,7 @@ class BatchInsertClient(Utils, BatchRetrieveClient, APIClient, Chunker):
             ):
 
                 # Get completed documents
-                log_json = self.get_all_documents(logging_dataset_id)
+                log_json = self._get_all_documents(logging_dataset_id)
                 completed_documents_list = [i["_id"] for i in log_json]
 
                 # Get incomplete documents from raw collection