+++ relevanceai/dataset_api/dataset.py
@@ -195,8 +195,9 @@ class Series(BatchAPIClient):
             Sort in ascending order.
         bins : int, optional
             Groups categories into 'bins'. These bins are good for representing groups within continuous series
+
         Returns
-        -------
+        ----------
         Series
         """
         schema = self.datasets.schema(self.dataset_id)
@@ -496,7 +497,7 @@ class Read(BatchAPIClient):
         elif output_format == "pandas":
             return pd.DataFrame.from_dict(docs, orient="records")
 
-    def all(
+    def get_all_documents(
         self,
         chunk_size: int = 1000,
         filters: List = [],
@@ -510,7 +511,7 @@ class Read(BatchAPIClient):
         Retrieve all documents with filters. Filter is used to retrieve documents that match the conditions set in a filter query. This is used in advance search to filter the documents that are searched. For more details see documents.get_where.
 
         Parameters
-        ----------
+        ------------
         chunk_size : list
             Number of documents to retrieve per retrieval
         include_vector: bool
@@ -521,9 +522,20 @@ class Read(BatchAPIClient):
             Query for filtering the search results
         select_fields : list
             Fields to include in the search results, empty array/list means all fields.
+
+        Example
+        ----------
+
+        .. code-block
+
+            from relevanceai import Client
+            client = Client()
+            df = client.Dataset("sample")
+            docs = df.get_all_documents()
+
         """
 
-        return self.get_all_documents(
+        return self._get_all_documents(
             dataset_id=self.dataset_id,
             chunk_size=chunk_size,
             filters=filters,
@@ -533,7 +545,9 @@ class Read(BatchAPIClient):
             show_progress_bar=show_progress_bar,
         )
 
-    def get(self, document_ids: Union[List, str], include_vector: bool = True):
+    def get_documents_by_ids(
+        self, document_ids: Union[List, str], include_vector: bool = True
+    ):
         """
         Retrieve a document by its ID ("_id" field). This will retrieve the document faster than a filter applied on the "_id" field.
 
@@ -549,7 +563,7 @@ class Read(BatchAPIClient):
         >>> from relevanceai import Client, Dataset
         >>> client = Client()
         >>> df = client.Dataset("sample_dataset")
-        >>> df.get("sample_id", include_vector=False)
+        >>> df.get_documents_by_ids(["sample_id"], include_vector=False)
 
         """
         if isinstance(document_ids, str):
@@ -660,6 +674,126 @@ class Write(Read):
             overwrite=overwrite,
         )
         return centroids
+
+    def insert_documents(  # type: ignore
+        self,
+        documents: list,
+        bulk_fn: Callable = None,
+        max_workers: int = 8,
+        retry_chunk_mult: float = 0.5,
+        show_progress_bar: bool = False,
+        chunksize: int = 0,
+        use_json_encoder: bool = True,
+        *args,
+        **kwargs,
+    ):
+
+        """
+        Insert a list of documents with multi-threading automatically enabled.
+
+        - When inserting the document you can optionally specify your own id for a document by using the field name "_id", if not specified a random id is assigned.
+        - When inserting or specifying vectors in a document use the suffix (ends with) "_vector_" for the field name. e.g. "product_description_vector_".
+        - When inserting or specifying chunks in a document the suffix (ends with) "_chunk_" for the field name. e.g. "products_chunk_".
+        - When inserting or specifying chunk vectors in a document's chunks use the suffix (ends with) "_chunkvector_" for the field name. e.g. "products_chunk_.product_description_chunkvector_".
+
+        Documentation can be found here: https://ingest-api-dev-aueast.relevance.ai/latest/documentation#operation/InsertEncode
+
+        Parameters
+        ----------
+        documents: list
+            A list of documents. Document is a JSON-like data that we store our metadata and vectors with. For specifying id of the document use the field '_id', for specifying vector field use the suffix of '_vector_'
+        bulk_fn : callable
+            Function to apply to documents before uploading
+        max_workers : int
+            Number of workers active for multi-threading
+        retry_chunk_mult: int
+            Multiplier to apply to chunksize if upload fails
+        chunksize : int
+            Number of documents to upload per worker. If None, it will default to the size specified in config.upload.target_chunk_mb
+        use_json_encoder : bool
+            Whether to automatically convert documents to json encodable format
+
+        Example
+        --------
+
+        >>> from relevanceai import Client
+        >>> client = Client()
+        >>> df = client.Dataset("sample_dataset")
+        >>> documents = [{"_id": "10", "value": 5}, {"_id": "332", "value": 10}]
+        >>> df.insert_documents(documents)
+
+        """
+        return self._insert_documents(  # type: ignore
+            dataset_id=self.dataset_id,
+            docs=documents,
+            bulk_fn=bulk_fn,
+            max_workers=max_workers,
+            retry_chunk_mult=retry_chunk_mult,
+            show_progress_bar=show_progress_bar,
+            chunksize=chunksize,
+            use_json_encoder=use_json_encoder,
+            *args,
+            **kwargs,
+        )
+
+    def insert_csv(  # type: ignore
+        self,
+        filepath_or_buffer,
+        chunksize: int = 10000,
+        max_workers: int = 8,
+        retry_chunk_mult: float = 0.5,
+        show_progress_bar: bool = False,
+        index_col: int = None,
+        csv_args: dict = {},
+        col_for_id: str = None,
+        auto_generate_id: bool = True,
+    ):
+
+        """
+        Insert data from csv file
+
+        Parameters
+        ----------
+        dataset_id : string
+            Unique name of dataset
+        filepath_or_buffer :
+            Any valid string path is acceptable. The string could be a URL. Valid URL schemes include http, ftp, s3, gs, and file.
+        chunksize : int
+            Number of lines to read from csv per iteration
+        max_workers : int
+            Number of workers active for multi-threading
+        retry_chunk_mult: int
+            Multiplier to apply to chunksize if upload fails
+        csv_args : dict
+            Optional arguments to use when reading in csv. For more info, see https://pandas.pydata.org/docs/reference/api/pandas.read_csv.html
+        index_col : None
+            Optional argument to specify if there is an index column to be skipped (e.g. index_col = 0)
+        col_for_id : str
+            Optional argument to use when a specific field is supposed to be used as the unique identifier ('_id')
+        auto_generate_id: bool = True
+            Automatically generateds UUID if auto_generate_id is True and if the '_id' field does not exist
+
+        Example
+        ---------
+        >>> from relevanceai import Client
+        >>> client = Client()
+        >>> df = client.Dataset("sample_dataset")
+        >>> csv_filename = "temp.csv"
+        >>> df.insert_csv(csv_filename)
+
+        """
+        return self._insert_csv(
+            dataset_id=self.dataset_id,
+            filepath_or_buffer=filepath_or_buffer,
+            chunksize=chunksize,
+            max_workers=max_workers,
+            retry_chunk_mult=retry_chunk_mult,
+            show_progress_bar=show_progress_bar,
+            index_col=index_col,
+            csv_args=csv_args,
+            col_for_id=col_for_id,
+            auto_generate_id=auto_generate_id,
+        )
 
     def apply(
         self,
@@ -951,7 +1085,7 @@ class Export(Read):
         kwargs: Optional
             see client.get_all_documents() for extra args
         """
-        documents = self.get_all_documents(self.dataset_id, **kwargs)
+        documents = self.get_all_documents(**kwargs)
         df = pd.DataFrame(documents)
         df.to_csv(filename)
 
@@ -968,7 +1102,7 @@ class Export(Read):
         list of documents in dictionary format
         """
         if orient == "records":
-            return self.get_all_documents(self.dataset_id)
+            return self.get_all_documents()
         else:
             raise NotImplementedError