+++ relevanceai/vector_tools/cluster.py
@@ -548,7 +548,7 @@ class Cluster(ClusterEvaluate, BatchAPIClient, ClusterBase):
         self.logger.warning(
             "Retrieving documents... This can take a while if the dataset is large."
         )
-        docs = self.get_all_documents(
+        docs = self._get_all_documents(
             dataset_id=dataset_id, filters=filters, select_fields=vector_fields
         )
 
@@ -683,7 +683,7 @@ class Cluster(ClusterEvaluate, BatchAPIClient, ClusterBase):
                 ".".join([cluster_field, vector_fields[0], alias])
             )
         # load the documents
-        docs = self.get_all_documents(
+        docs = self._get_all_documents(
             dataset_id=dataset_id, filters=filters, select_fields=vector_fields
         )